Refactor OLED and e-paper blocks from communicate.js to oled.js.
